requirements.txt
Django==5.0.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-spectacular==0.27.2
django-simple-history==3.4.0
django-mptt==0.15.0
django-storages==1.14.2
boto3==1.34.144
psycopg[binary]==3.1.19
pgvector==0.2.5
celery==5.3.4
redis==5.0.7
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.4.0
factory-boy==3.3.0
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
mypy==1.10.1
django-stubs==5.0.2
djangorestframework-stubs==3.14.5
